refactor(ton): extract TonApi DTOs into TonApiDto.kt
Splits the wire-format data classes out of TonApi.kt into a sibling
file in the same package. Private DTOs become internal so they remain
accessible from TonApiImpl; public DTOs keep their visibility.
